+++ test/model_creation_tests.jl
@@ -36,12 +36,60 @@ using Test, Agents, Random
     @test Fisher <: AbstractHuman
     @test :fish_per_day ∈ fieldnames(Fisher)
 
+    @agent Agent9 NoSpaceAgent begin
+        f1::Int = 40
+        f2::Int
+        f3::Float64 = 3.0
+    end
     agent_kwdef = Agent9(id = 1, f2 = 10)
     values = (1, 40, 10, 3.0)
     @test all(getfield(agent_kwdef, n) == v for (n, v) in zip(fieldnames(Agent9), values))
     agent_kwdef = Agent9(1, 20, 10, 4.0)
     values = (1, 20, 10, 4.0)
     @test all(getfield(agent_kwdef, n) == v for (n, v) in zip(fieldnames(Agent9), values))
+
+    @agent Agent10 NoSpaceAgent begin
+        f1::Int
+        f2::Int
+        f3::Float64
+        constants = (:f2, )
+    end
+    agent_consts = Agent10(1, 2, 10, 5.0)
+    values = (1, 2, 10, 5.0)
+    @test all(getfield(agent_consts, n) == v for (n, v) in zip(fieldnames(Agent10), values))
+    agent_consts.f1 = 5
+    @test agent_consts.f1 == 5
+    @test_throws ErrorException agent_consts.f2 = 5
+
+    @agent Agent11 NoSpaceAgent begin
+        f1::Int
+        f2
+        f3::Float64
+        constants = (:f1, :f2)
+    end
+    agent_consts = Agent11(1, 2, 10, 5.0)
+    values = (1, 2, 10, 5.0)
+    @test all(getfield(agent_consts, n) == v for (n, v) in zip(fieldnames(Agent11), values))
+    agent_consts.f3 = 2.0
+    @test agent_consts.f3 == 2.0
+    @test_throws ErrorException agent_consts.f1 = 5
+    @test_throws ErrorException agent_consts.f2 = 5
+
+    @agent Agent12 Agent11 begin
+        f4
+        f5::Float64
+        constants = (:f4, )
+    end
+    agent_consts = Agent12(1, 2, 10, 5.0, true, 3.0)
+    values = (1, 2, 10, 5.0, true, 3.0)
+    @test all(getfield(agent_consts, n) == v for (n, v) in zip(fieldnames(Agent11), values))
+    agent_consts.f3 = 2.0
+    @test agent_consts.f3 == 2.0
+    agent_consts.f5 = 4.0
+    @test agent_consts.f5 == 4.0
+    @test_throws ErrorException agent_consts.f1 = 5
+    @test_throws ErrorException agent_consts.f2 = 5
+    @test_throws ErrorException agent_consts.f4 = false
 end
 
 
@@ -146,3 +194,4 @@ end
     ) ABM(Union{NoSpaceAgent,ValidAgent})
     @test_throws ArgumentError ABM(Union{NoSpaceAgent,BadAgent}; warn = false)
 end
+