+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # main
 
 - The `@agent` macro now supports fields with default and const values (through the special `constants` field). Since now the macro supports these features, using `@agent` is the only supported way to create agent types for Agents.jl.
-- The `add_agent!` function supports adding agent propagating keywords arguments.
+- The `add_agent!` function supports adding an agent propagating keywords arguments.
 
 # v5.17